fix(stella-model,stella-core): a sustained 529 brownout parks instead of aborting the turn
529 classified as Transport (#2742)
-----------------------------------
`classify_http_status` mapped every 5xx — including 529, which Anthropic
and Z.ai use for "overloaded" — to retryable `Transport`, while the #2677
parked recovery engages only for `RateLimited`. So a sustained brownout
exhausted the ~16s inline ladder and aborted the turn with wall-clock
budget still on the table: the exact shape #2667 measured for 429s. Both
halves were reproduced before the fix.

Sketch (a), a distinct `ProviderError::Overloaded { message,
retry_after_ms }`, not (b) a predicate flag. The repo's own test for when
a variant is required is in `stella-runtime/src/error.rs`: a caller that
must BRANCH needs a variant, not a string or a flag. The retry driver
decides park-vs-abort, which is a different decision from
retry-vs-abort — and a flag on `Transport` would leave the error's class
unchanged, so every consumer that already branches on the taxonomy
(`stella-serve`'s `ProviderErrorWire`, `live_smoke`'s deliberately
exhaustive `classify`) would have to learn to read it or silently treat
an overload as a generic fault. It is also the precedent set twice in
this enum: `ContextOverflow` and `OutputBudgetExceeded` were both split
out of `Terminal` for the identical reason, and say so.

The wire cost the issue warns about was paid, not dodged:
`ProviderErrorWire` gained an arm, both `From` impls, `make wire-schema`
regenerated and committed, and the round-trip test extended.

L-M4 holds: `RetryPolicy::deterministic()` never parks, pinned by a test.
New logic went to `driver/rate_limit.rs` and `retry.rs`, never into
`driver.rs`, which is a god file at its ceiling.

max_tokens 128000 (#3757) — NOT CLOSED, comment only
-----------------------------------------------------
I asked for the ceiling to drop to 64000. That was wrong, and the fix
declines it on evidence rather than complying.

The issue's premise — "a default nobody chose" — is inverted. This row
read 64000 until 2026-08-03, when #1290 *deliberately raised* it, citing
two independent sources (Anthropic's `GET /v1/models` reports
`"max_tokens": 128000`; OpenRouter reports `top_provider.
max_completion_tokens = 128000`) and recording that 64000 "was never the
model's ceiling — it was where Claude Code's steps were measured
stopping, which is a fact about the comparator's posture, not about
Sonnet 5". 64000 is the unchosen number, and it was already fixed once.

The hazard is also already mitigated:
`stella_core::driver::output_budget_recovery` classifies the gateway's
402, clamps below the stated affordable figure, and re-runs the step,
with `SessionOutputCeilings` (#3307) carrying the clamp across turns. And
"299 output tokens per call" cannot decide a ceiling at all — a ceiling
bounds the longest answer, not the median.

So this takes the issue's OTHER definition-of-done branch, which it lists
as equally acceptable: the 128000 value now carries a comment saying why
it is worth the refusal risk. Zero behaviour change — Stella still sends
128000 — so `Refs`, not `Closes`. The branch-1 question is a maintainer's
call and is put to one in the PR.

That comment was trimmed from 34 lines to 20 after review, for two
reasons. It restated `output_budget_recovery`'s module doc instead of
pointing at it, and it hard-coded "once-per-eight-turns" — a second copy
of `REPROBE_TURNS: u32 = 8` living in a crate that cannot see the
constant, which is how the numbers in this repository have died before.
`catalog.rs` is also 1423 lines against a hard 1500 with no baseline
entry; filed as #3862, because the next few model seeds cross it.

Verification
------------
cargo test -p stella-model, -p stella-core, -p stella-cli: no failures.
check-wire-schema OK. clippy over model/core/protocol/serve clean.
cargo fmt --check clean. Eleven guards incl. typed-errors and
diagnostic-codes: OK. Both ratchet baselines untouched.

Closes #2742
Refs #3757, #2677, #2667, #3307, #1290, #3862
